Correct MySQL string literal escaping
